requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core


langgraph
langgraph-prebuilt


langserve[all]
fastapi
uvicorn
python-dotenv


faiss-cpu
pydantic
pydantic-settings
pypdf


mcp[cli]
markdown

pytest
pytest-asyncio
httpx

streamlit